+++ CHANGES.md
@@ -1,3 +1,13 @@
+# 2024-01-06 (5.24.1)
+
+* Fix bug in exporter where a loop through dataset tables was prematurely
+  left when a tables has not export data.
+
+# 2024-01-06 (5.24.0)
+
+* Do not write an empty export file if no columns are selected.
+  Also fix the export to only use active records for jsonlines.
+
 # 2024-01-31 (5.23.4)
 
 * Bugfix in _is_valid_sql
@@ -9,7 +19,7 @@
 # 2024-01-23 (5.23.2)
 
 * Fix to the _get_scopes to return the correct scopes for both dataset, table
-  and table fields. 
+  and table fields.
 
 # 2024-01-22 (5.23.1)